Squashed 'src/secp256k1/' changes from 367649e587..9224bf7688
9224bf7688 TEST2

git-subtree-dir: src/secp256k1
git-subtree-split: 9224bf7688347f6e7d4aa0ac2e2d7096c4073a2f

+++ src/hash_impl.h
@@ -127,7 +127,7 @@ static void secp256k1_sha256_transform_impl(uint32_t* s, const unsigned char* bu
     s[7] += h;
 }
 
-static void secp256k1_sha256_transform(uint32_t* s, const unsigned char* buf, uint32_t n) {
+static void secp256k1_sha256_transform(uint32_t* s, const unsigned char* buf, size_t n) {
     while (n--) secp256k1_sha256_transform_impl(s, buf);
 }
